requirements.txt
apache-airflow[otel]==2.10.3
apache-airflow[statsd]==2.10.3